// src/models/Pago.ts
/**
 *
 * @export
 * @interface Pago
 */
export interface Pago {
    /**
     *
     * @type {number}
     * @memberof Pago
     */
    Id: number;
  
    /**
     *
     * @type {string}
     * @memberof Pago
     */
    Nombre: string;
  
    /**
     *
     * @type {string}
     * @memberof Pago
     */
    Apellido: string;
  
    /**
     *
     * @type {number}
     * @memberof Pago
     */
    Total: number;
  
    /**
     *
     * @type {string}
     * @memberof Pago
     */
    Pagado: string;
  }
  